// @flywave/flywave-topo/src/bspline/bspline-curve.d.ts
export declare abstract class BSplineCurve3dBase extends CurvePrimitive {
    readonly curvePrimitiveType = "bsplineCurve";
    protected _bcurve: BSpline1dNd;
    private _definitionData?;
    set definitionData(data: any);
    get definitionData(): any;
    protected constructor(poleDimension: number, numPoles: number, order: number, knots: KnotVector);
    get degree(): number;
    get order(): number;
    get numSpan(): number;
    get numPoles(): number;
    copyKnots(includeExtraEndKnot: boolean): number[];
    setWrappable(value: BSplineWrapMode): void;
    abstract evaluatePointInSpan(spanIndex: number, spanFraction: number, result?: Point3d): Point3d;
    abstract evaluatePointAndDerivativeInSpan(spanIndex: number, spanFraction: number, result?: Ray3d): Ray3d;
    abstract knotToPoint(knot: number, result?: Point3d): Point3d;
    abstract knotToPointAndDerivative(knot: number, result?: Ray3d): Ray3d;
    abstract knotToPointAnd2Derivatives(knot: number, result?: Plane3dByOriginAndVectors): Plane3dByOriginAndVectors;
    fractionToPoint(fraction: number, result?: Point3d): Point3d;
    fractionToPointAndDerivative(fraction: number, result?: Ray3d): Ray3d;
    fractionToPointAnd2Derivatives(fraction: number, result?: Plane3dByOriginAndVectors): Plane3dByOriginAndVectors;
    startPoint(): Point3d;
    endPoint(): Point3d;
    reverseInPlace(): void;
    collectBezierSpans(prefer3dH: boolean): BezierCurveBase[];
    abstract getSaturatedBezierSpan3dOr3dH(spanIndex: number, prefer3dH: boolean, result?: BezierCurveBase): BezierCurveBase | undefined;
    abstract getPolePoint4d(poleIndex: number, result?: Point4d): Point4d | undefined;
    abstract getPolePoint3d(poleIndex: number, result?: Point3d): Point3d | undefined;
    poleIndexToDataIndex(poleIndex: number): number | undefined;
    closestPoint(spacePoint: Point3d, _extend: boolean): CurveLocationDetail | undefined;
    abstract clone(): BSplineCurve3dBase;
    cloneTransformed(transform: Transform): BSplineCurve3dBase;
    clonePartialCurve(fractionA: number, fractionB: number): BSplineCurve3dBase;
    appendPlaneIntersectionPoints(plane: PlaneAltitudeEvaluator, result: CurveLocationDetail[]): number;
    constructOffsetXY(offsetDistanceOrOptions: number | OffsetOptions): CurvePrimitive | CurvePrimitive[] | undefined;
    projectedParameterRange(ray: Vector3d | Ray3d, lowHigh?: Range1d): Range1d | undefined;
}
export declare class BSplineCurve3d extends BSplineCurve3dBase {
    private _workBezier?;
    private initializeWorkBezier;
    isSameGeometryClass(other: any): boolean;
    tryTransformInPlace(transform: Transform): boolean;
    getPolePoint3d(poleIndex: number, result?: Point3d): Point3d | undefined;
    getPolePoint4d(poleIndex: number, result?: Point4d): Point4d | undefined;
    spanFractionToKnot(span: number, localFraction: number): number;
    private constructor();
    copyPoints(): any[];
    copyPointsFloat64Array(): Float64Array;
    copyKnots(includeExtraEndKnot: boolean): number[];
    static createUniformKnots(poles: Point3d[] | Float64Array | GrowableXYZArray, order: number): BSplineCurve3d | undefined;
    static createPeriodicUniformKnots(poles: Point3d[] | Float64Array | GrowableXYZArray, order: number): BSplineCurve3d | undefined;
    static createFromInterpolationCurve3dOptions(options: InterpolationCurve3dOptions): BSplineCurve3d | undefined;
    static createFromAkimaCurve3dOptions(options: AkimaCurve3dOptions): BSplineCurve3d | undefined;
    static create(poleArray: Float64Array | Point3d[], knotArray: Float64Array | number[], order: number): BSplineCurve3d | undefined;
    clone(): BSplineCurve3d;
    evaluatePointInSpan(spanIndex: number, spanFraction: number): Point3d;
    evaluatePointAndDerivativeInSpan(spanIndex: number, spanFraction: number): Ray3d;
    knotToPoint(u: number, result?: Point3d): Point3d;
    knotToPointAndDerivative(u: number, result?: Ray3d): Ray3d;
    knotToPointAnd2Derivatives(u: number, result?: Plane3dByOriginAndVectors): Plane3dByOriginAndVectors;
    isAlmostEqual(other: any): boolean;
    isInPlane(plane: Plane3dByOriginAndUnitNormal): boolean;
    quickLength(): number;
    emitStrokableParts(handler: IStrokeHandler, options?: StrokeOptions): void;
    computeStrokeCountForOptions(options?: StrokeOptions): number;
    computeAndAttachRecursiveStrokeCounts(options?: StrokeOptions, parentStrokeMap?: StrokeCountMap): void;
    emitStrokes(dest: LineString3d, options?: StrokeOptions): void;
    get isClosable(): BSplineWrapMode;
    getSaturatedBezierSpan3dOr3dH(spanIndex: number, prefer3dH: boolean, result?: BezierCurveBase): BezierCurveBase | undefined;
    getSaturatedBezierSpan3d(spanIndex: number, result?: BezierCurveBase): BezierCurveBase | undefined;
    getSaturatedBezierSpan3dH(spanIndex: number, result?: BezierCurveBase): BezierCurve3dH | undefined;
    setWrappable(value: BSplineWrapMode): void;
    dispatchToGeometryHandler(handler: GeometryHandler): any;
    extendRange(rangeToExtend: Range3d, transform?: Transform): void;
}